Increased the maximum grain radius
With variable grain radius, the vision filter uses an estimation of the
maximum expected radius to limit the scanned cell area. But if the
actual radius of a grain is too large, the grain shape is clipped to
square. The previous value (2x std dev) was a bit too small and too
much grains appeared clipped with a large std dev, affecting not only
the grain shape but also the resulting average luminance, when grains
are small.