Added a draft mode (fast rendering)
Works only for a subset of the parameter combinations.
Not very accurate at the moment because it removes one layer of
randomness (geometric grain modeling) and only keeps the random
quantity of grains depending on the luminance.

To improve the draft mode, we should model how random grain
overlapping maps to luminance changes.